fix(algo): align order cancellations
Move Spot and Futures Algo cancellation payloads to signed DELETE query parameters, enforce the receive-window ceiling, align canonical identifiers and 64-bit responses, and add deterministic regression coverage.

+++ Binance.Api/Algo/Responses/BinanceAlgoResult.cs
@@ -9,7 +9,7 @@ public record BinanceAlgoResult
     /// Result code
     /// </summary>
     [JsonProperty("code")]
-    public int Code { get; set; }
+    public long Code { get; set; }
 
     /// <summary>
     /// Message
